testing DOM manipulation functions
